update(sm): reformat sm resources and header added
